Move config.js to src as well.
